backend/requirements.txt
# Django Core
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# CORS
django-cors-headers==4.3.1

# OpenAPI/Swagger
drf-spectacular==0.27.0

# Environment variables
python-decouple==3.8

# Utilities
pytz==2023.3
